Consolodate navbars
