+++ client/src/components/signup/Signup.js
@@ -86,8 +86,10 @@ const Signup = () => {
   return (
     <div class="max-w-4xl mx-auto">
       <div class="bg-white shadow-md border border-gray-200 rounded-lg max-w-3xl p-4 sm:p-8 lg:p-12 dark:bg-white dark:border-gray-300">
-        <form class="space-y-6" onSubmit={handleSubmit}>
-          <h2 className="text-3xl text-grey-700">Sign up</h2>
+        <form class="space-y-1" onSubmit={handleSubmit}>
+          <h2 className="text-3xl text-grey-700 dark: text-gray-800 mb-4">
+            Sign up
+          </h2>
           <p ref={errRef} className={errorMsg ? "text-red-600" : "hidden"}>
             {errorMsg}
           </p>
@@ -200,12 +202,14 @@ const Signup = () => {
               value={confirmPassword}
             />
           </div>
-          <button
-            type="submit"
-            class="w-full text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-gradient-to-r from-purple-500 to-pink-500 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
-          >
-            Login to your account
-          </button>
+          <div className="">
+            <button
+              type="submit"
+              class="w-full text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-gradient-to-r from-purple-500 to-pink-500 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+            >
+              Create your account
+            </button>
+          </div>
           <div class="text-sm font-medium text-gray-500 dark:text-gray-900">
             Registered?{" "}
             <Link
