+++ client/src/pages/Signerino.js
@@ -1,5 +1,5 @@
 import React from "react";
-import UnauthedNavbar from "../components/navbar/UnauthedNavbar";
+import Navbar from "../components/navbar/Navbar";
 import Signup from "../components/signup/Signup";
 import { useAuth } from "../context/AuthProvider";
 import { Navigate } from "react-router-dom";
@@ -14,7 +14,7 @@ const Signerino = () => {
     <div className="">
       <div>
         <div>
-          <UnauthedNavbar></UnauthedNavbar>
+          <Navbar></Navbar>
         </div>
         <div className="relative flex items-center justify-center">
           <div className="w-full md:w-2/3 self-center max-w-xl lg:max-w-2xl h-full">
